search by selection fixed, begining of discogs added
